+++ packages/create-web/templates/library/test-setup.js
@@ -1,0 +1,4 @@
+import { otfwcPath } from "@opentf/web-compiler";
+
+process.env.OTFWC_BIN ??= otfwcPath();
+await import("@opentf/web-test/setup");